Address Andres' comments.
